+++ installers/bundles/__init__.py
@@ -1,0 +1,1 @@
+"""Bundle-oriented installer domains."""